backend/app/schemas/tenant.py
from datetime import datetime
from uuid import UUID
from pydantic import BaseModel, ConfigDict

class TenantBase(BaseModel):
    name: str

class TenantCreate(TenantBase):
    pass

class TenantUpdate(TenantBase):
    pass

class Tenant(TenantBase):
    id: UUID
    created_at: datetime
    updated_at: datetime

    model_config = ConfigDict(from_attributes=True)
